move converters test to tests module
