+++ app/pom.xml
@@ -139,11 +139,6 @@
         </dependency>
         <dependency>
             <groupId>io.apicurio</groupId>
-            <artifactId>apicurio-registry-utils-converter</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>io.apicurio</groupId>
             <artifactId>apicurio-registry-maven-plugin</artifactId>
             <type>maven-plugin</type>
             <scope>test</scope>
