+++ tests/pom.xml
@@ -148,6 +148,11 @@
             <artifactId>apicurio-registry-utils-serde</artifactId>
             <scope>compile</scope>
         </dependency>
+        <dependency>
+            <groupId>io.apicurio</groupId>
+            <artifactId>apicurio-registry-utils-converter</artifactId>
+            <scope>compile</scope>
+        </dependency>
     </dependencies>
 
     <build>
